Define integration base class
